Source code and basic circuit diagram
This upload is the first pass at making the circuit and source code available to the public. It comes with no guarantee of suitability for any purpose. The circuit diagrams are not necessarily free from errors. Use your own judgement and experience in wiring any circuit. Specifically, the values of resistors should be checked as I can't remember whether the fritzing file was corrected to the values used in the prototype.